pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-sync"
version = "0.1.0"
description = "Sync Claude.ai project knowledge to local storage"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Eric Downes", email = "eric@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "websocket-client>=1.6.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "rich>=13.0.0",  # For nice CLI output
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "responses>=0.23.0",
]

[project.scripts]
claude-sync = "src.cli.main:cli"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = [
    ".",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C90", "UP", "S"]
ignore = ["E501", "S101"]  # E501: line too long, S101: use of assert

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true